blog/requirements.txt
Django>=5.0
Pillow
psycopg2-binary>=2.9
django-filter>=23.0
djangorestframework>=3.15
djangorestframework-simplejwt==5.5.1
drf-yasg>=1.21